requirements.txt
gutils @ https://github.com/giussepi/gutils/tarball/master
tensorboardX>=2.1
tensorboard>=2.4.0
torchmetrics>=0.7.2
logzero>=1.6.3

# emacs as python editor
autopep8>=1.5.4
ipython>=7.16.1
jedi>=0.17.2
pyflakes>=2.2.0
Pygments>=2.7.2
pylint>=2.6.0
